split PsUpdateStream out of models
