+++ osidb/dmodels/__init__.py
@@ -1,4 +1,5 @@
 from .ps_module import PsModule
 from .ps_product import PsProduct
+from .ps_update_stream import PsUpdateStream
 
-__all__ = ("PsModule", "PsProduct")
+__all__ = ("PsModule", "PsProduct", "PsUpdateStream")
